+++ lib/src/utilities/utilities.dart
@@ -4,3 +4,4 @@ export '../mixins/naked_mixins.dart';
 export 'positioning.dart';
 export 'widget_state_snapshot.dart';
 export 'widget_states_extension.dart';
+export 'naked_focusable_detector.dart';